Merge branch 'develop' into feature/multiresvolume

* develop: (25 commits)
  Compile fix
  Improvements from code review
  Update Ghoul repository Adapt to changed signature of ghoul_assert macro
  Update include finding script
  Various cleanup Add script that counts the number of includes
  Move the PropertyOwner name specification into the constructor and adjust accordingly
  Don't assert if the focus node is not found if there are no nodes at all
  Some cleanup of ModuleEngine Some cleanup of SettingsEngine Some cleanup of SyncEngine
  - Some cleanup of LuaConsole   - Moved dedicated Lua commands into a visible property   - Disallow non-ASCII characters - Remove unused KeyboardController - Remove unused DeviceIdentifier - Some cleanup of LogFactory   - Move and use documentation into the class
  Change to correct include guard name
  Adapt to new interfaces
  Cleanup
  - Remove documentation.h include and replace with forward declaration - Move Documentation struct into documentation namespace
  Some cleanup of ConfigurationManager
  Update branch to work with changes in develop
  Jenkins compilation fix
  Clean up volume rendering code
  Fix inside volume rendering crash on certain platforms
  Resolve some A-buffer related shader issues
  Compile fixes
  ...

# Conflicts:
#	ext/ghoul

+++ modules/volume/volumeutils.h
@@ -30,8 +30,8 @@
 namespace openspace {
 namespace volumeutils {
 
-size_t coordsToIndex(const glm::vec3& coords, const glm::ivec3& dimensions);
-glm::vec3 indexToCoords(size_t index, const glm::ivec3& dimensions);
+size_t coordsToIndex(const glm::uvec3& coords, const glm::uvec3& dimensions);
+glm::uvec3 indexToCoords(size_t index, const glm::uvec3& dimensions);
 
 } // namespace volumeutils
